added jwt authentication

+++ backend/controller/goalController.js
@@ -1,13 +1,19 @@
 const asyncHandler = require('express-async-handler');
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel');
 
 // @description:     Get Goals
 // @route:          GET /api/goals
 // @access          Private
 
 const getGoals = asyncHandler(async (req,res) => {
-    const goals = await Goal.find();
-    res.status(200).json(goals)
+    const goals = await Goal.find({
+        user:req.user.id
+    });
+    
+        res.status(200).json(goals)
+    
+    
 })
 
 // @description:     Set Goals
@@ -20,6 +26,7 @@ const setGoals = asyncHandler( async (req,res) => {
         throw new Error('please add a text field')
     }
      const goal = await Goal.create({
+        user:req.user.id,
         text:req.body.text
      })
     console.log(req.body);
@@ -32,10 +39,19 @@ const setGoals = asyncHandler( async (req,res) => {
 
 const updateGoals = asyncHandler( async (req,res) => {
     const goal = await Goal.findById(req.params.id);
-
+    const user = await User.findById(req.user.id);
     if(!goal) {
         res.status(400);
         throw new Error('goal not found');;
+    }
+    if (!user) {
+        res.status(401);
+        throw new Error('Invalid user');
+    }
+
+    if(goal.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error('user not authorized');
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -50,10 +66,22 @@ const updateGoals = asyncHandler( async (req,res) => {
 
 const deleteGoals = asyncHandler( async (req,res) => {
     const goal = await Goal.findById(req.params.id);
+    const user = await User.findById(req.user.id);
     if(!goal) {
         res.status(400);
         throw new Error('goal not found');
     }
+
+    if (!user) {
+        res.status(401);
+        throw new Error('Invalid user');
+    }
+
+    if(goal.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error('user not authorized');
+    }
+
 
     const deletedGoal = await Goal.findByIdAndDelete(req.params.id);
     res.status(200).json(deletedGoal);
